Add files via upload
This CCAMA is completely rid of compile errors and is finished up to line 169 of the MATLAB code. There are a few things missing in this chunk of the code such as the ability to determine if the function has 7 or  8 arguments; thus, in this code right now CCAMA only supports 7 arguments thus uses default options. It is also missing the printf function and timer function.